update vector store op

+++ experiencemaker/vector_store/es_vector_store.py
@@ -16,26 +16,23 @@ from experiencemaker.vector_store.base_vector_store import BaseVectorStore
 class EsVectorStore(BaseVectorStore):
     hosts: str | List[str] = Field(default_factory=lambda: os.getenv("ES_HOSTS", "http://localhost:9200"))
     basic_auth: str | Tuple[str, str] | None = Field(default=None)
-    bulk_chunk_size: int = Field(default=512)
     retrieve_filters: List[dict] = []
     _client: Elasticsearch = PrivateAttr()
 
     @model_validator(mode="after")
     def init_client(self):
         if isinstance(self.hosts, str):
-            hosts = [self.hosts]
-        else:
-            hosts = self.hosts
-        self._client = Elasticsearch(hosts=hosts, basic_auth=self.basic_auth)
+            self.hosts = [self.hosts]
+        self._client = Elasticsearch(hosts=self.hosts, basic_auth=self.basic_auth)
         return self
 
     def exist_workspace(self, workspace_id: str, **kwargs) -> bool:
         return self._client.indices.exists(index=workspace_id)
 
-    def _delete_workspace(self, workspace_id: str, **kwargs):
-        self._client.indices.delete(index=workspace_id, **kwargs)
+    def delete_workspace(self, workspace_id: str, **kwargs):
+        return self._client.indices.delete(index=workspace_id, **kwargs)
 
-    def _create_workspace(self, workspace_id: str, **kwargs):
+    def create_workspace(self, workspace_id: str, **kwargs):
         body = {
             "mappings": {
                 "properties": {
@@ -49,14 +46,10 @@ class EsVectorStore(BaseVectorStore):
                 }
             }
         }
-
         return self._client.indices.create(index=workspace_id, body=body)
 
-    def _iter_workspace_nodes(self, workspace_id: str, max_size: int = 10000, **kwargs) -> Iterable[VectorNode]:
-        response = self._client.search(index=workspace_id,
-                                       body={"query": {"match_all": {}}, "size": max_size},
-                                       scroll='5m')
-
+    def _iter_workspace_nodes(self, workspace_id: str, **kwargs) -> Iterable[VectorNode]:
+        response = self._client.search(index=workspace_id, body={"query": {"match_all": {}}})
         for doc in response['hits']['hits']:
             yield self.doc2node(doc)
 
@@ -70,26 +63,6 @@ class EsVectorStore(BaseVectorStore):
         if "_score" in doc:
             node.metadata["_score"] = doc["_score"] - 1
         return node
-
-    def exist_id(self, unique_id: str, workspace_id: str = None, **kwargs) -> bool:
-        response = self._client.exists(index=workspace_id, id=unique_id)
-        return response.body
-
-    def node2doc(self, node: VectorNode, add_op_type: bool = False) -> dict:
-        doc: dict = {
-            "_index": node.workspace_id,
-            "_id": node.unique_id,
-            "_source": {
-                "workspace_id": node.workspace_id,
-                "content": node.content,
-                "metadata": node.metadata,
-                "vector": node.vector
-            }
-        }
-
-        if add_op_type:
-            doc["_op_type"] = "update" if self.exist_id(node.unique_id, node.workspace_id) else "index",
-        return doc
 
     def add_term_filter(self, key: str, value):
         if key:
@@ -110,7 +83,7 @@ class EsVectorStore(BaseVectorStore):
         self.retrieve_filters.clear()
         return self
 
-    def retrieve_by_query(self, query: str, workspace_id: str, top_k: int = 1, **kwargs) -> List[VectorNode]:
+    def search(self, query: str, workspace_id: str, top_k: int = 1, **kwargs) -> List[VectorNode]:
         if not self.exist_workspace(workspace_id=workspace_id):
             logger.warning(f"workspace_id={workspace_id} is not exists!")
             return []
@@ -137,8 +110,10 @@ class EsVectorStore(BaseVectorStore):
         self.retrieve_filters.clear()
         return nodes
 
-    def insert(self, nodes: VectorNode | List[VectorNode], workspace_id: str, refresh: bool = True, **kwargs):
-        self.create_workspace(workspace_id=workspace_id)
+    def insert(self, nodes: VectorNode | List[VectorNode], workspace_id: str, refresh: bool = False, **kwargs):
+        if not self.exist_workspace(workspace_id=workspace_id):
+            self.create_workspace(workspace_id=workspace_id)
+
         if isinstance(nodes, VectorNode):
             nodes = [nodes]
 
@@ -146,38 +121,55 @@ class EsVectorStore(BaseVectorStore):
         not_embedded_nodes = [node for node in nodes if not node.vector]
         now_embedded_nodes = self.embedding_model.get_node_embeddings(not_embedded_nodes)
 
-        docs = [self.node2doc(node, False) for node in embedded_nodes + now_embedded_nodes]
-        status, error = bulk(self._client, docs, chunk_size=self.bulk_chunk_size, **kwargs)
-        logger.info(f"insert sample.size={len(nodes)} status={status} error={error}")
+        docs = [
+            {
+                "_op_type": "index",
+                "_index": node.workspace_id,
+                "_id": node.unique_id,
+                "_source": {
+                    "workspace_id": node.workspace_id,
+                    "content": node.content,
+                    "metadata": node.metadata,
+                    "vector": node.vector
+                }
+            } for node in embedded_nodes + now_embedded_nodes]
+        status, error = bulk(self._client, docs, chunk_size=self.batch_size, **kwargs)
+        logger.info(f"insert docs.size={len(docs)} status={status} error={error}")
 
         if refresh:
             self.refresh(workspace_id=workspace_id)
 
-    def update(self, nodes: VectorNode | List[VectorNode], workspace_id: str, refresh: bool = True, **kwargs):
-        self.create_workspace(workspace_id=workspace_id)
-        if isinstance(nodes, VectorNode):
-            nodes = [nodes]
+    def delete(self, node_ids: str | List[str], workspace_id: str, refresh: bool = False, **kwargs):
+        if not self.exist_workspace(workspace_id=workspace_id):
+            logger.warning(f"workspace_id={workspace_id} is not exists!")
+            return
 
-        nodes = self.embedding_model.get_node_embeddings(nodes)
-        docs = [self.node2doc(node, True) for node in nodes]
-        status, error = bulk(self._client, docs, chunk_size=self.bulk_chunk_size, **kwargs)
-        update_size = sum([1 if doc["_op_type"] == "update" else 0 for doc in docs])
-        insert_size = len(docs) - update_size
-        logger.info(f"update update_size={update_size} insert_size={insert_size} status={status} error={error}")
+        if isinstance(node_ids, str):
+            node_ids = [node_ids]
+
+        actions = [
+            {
+                "_op_type": "delete",
+                "_index": workspace_id,
+                "_id": node_id
+            } for node_id in node_ids]
+        status, error = bulk(self._client, actions, chunk_size=self.batch_size, **kwargs)
+        logger.info(f"delete actions.size={len(actions)} status={status} error={error}")
 
         if refresh:
             self.refresh(workspace_id=workspace_id)
+
 
 def main():
-    from experiencemaker.utils.util_function import load_env_keys
-    load_env_keys()
-    load_env_keys("../../.env")
+    from dotenv import load_dotenv
+    load_dotenv()
 
-    embedding_model = OpenAICompatibleEmbeddingModel(dimensions=1024)
+    embedding_model = OpenAICompatibleEmbeddingModel(dimensions=64, model_name="text-embedding-v4")
     workspace_id = "rag_nodes_index"
     hosts = "http://11.160.132.46:8200"
     es = EsVectorStore(hosts=hosts, embedding_model=embedding_model)
-    es.delete_workspace(workspace_id=workspace_id)
+    if es.exist_workspace(workspace_id=workspace_id):
+        es.delete_workspace(workspace_id=workspace_id)
     es.create_workspace(workspace_id=workspace_id)
 
     sample_nodes = [
@@ -215,13 +207,13 @@ def main():
 
     logger.info("=" * 20)
     results = es.add_term_filter(key="metadata.node_type", value="n1") \
-        .retrieve_by_query("What is AI?", top_k=5, workspace_id=workspace_id)
+        .search("What is AI?", top_k=5, workspace_id=workspace_id)
     for r in results:
         logger.info(r.model_dump(exclude={"vector"}))
     logger.info("=" * 20)
 
     logger.info("=" * 20)
-    results = es.retrieve_by_query("What is AI?", top_k=5, workspace_id=workspace_id)
+    results = es.search("What is AI?", top_k=5, workspace_id=workspace_id)
     for r in results:
         logger.info(r.model_dump(exclude={"vector"}))
     logger.info("=" * 20)
